pruebas unitarias todo el modulo usuarios CRUD

+++ src/main/java/com/meself/usuarios/meself_usuarios/controllers/UsuarioController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meself.usuarios.meself_usuarios.DTO.UsuarioDTO;
+import com.meself.usuarios.meself_usuarios.basic.Utils;
 import com.meself.usuarios.meself_usuarios.models.UsuarioEntity;
 import com.meself.usuarios.meself_usuarios.services.IUsuarioService;
 
@@ -21,9 +22,16 @@ public class UsuarioController {
 @Autowired
 IUsuarioService usuarioService;
 
+@Autowired
+Utils utils;
+
 
     @PostMapping("/createUsuario")
     public ResponseEntity<?> createUsuario(@RequestBody UsuarioDTO usuarioDTO){
+
+        if(utils.isObjectInvalid(usuarioDTO)){
+            return ResponseEntity.badRequest().body("Parametro necesario: usuario");
+         }
 
         return ResponseEntity.ok(usuarioService.createUsuarioEntity(usuarioDTO));
     }
@@ -31,19 +39,21 @@ IUsuarioService usuarioService;
     @GetMapping("/getUsuario/{usuarioId}")
     public ResponseEntity<?> getUserById(@PathVariable Long usuarioId){
 
+        if(utils.isObjectInvalid(usuarioId)){
+            return ResponseEntity.badRequest().body("Parametro necesario: usuarioId");
+         }
+
         UsuarioEntity usuarioEntity = usuarioService.getUsuarioEntityByUsuarioId(usuarioId);
 
         return ResponseEntity.ok(usuarioEntity);
     }
 
-    @GetMapping("/hello")
-    public String hello(){
-
-        return "Hola amios";
-    }
-
     @DeleteMapping("/deleteUsuario/{usuarioId}")
     public ResponseEntity<?> deleteUserById(@PathVariable Long usuarioId){
+
+        if(utils.isObjectInvalid(usuarioId)){
+           return ResponseEntity.badRequest().body("Parametro necesario: usuarioId");
+        }
 
          usuarioService.deleteUsuariosEntityById(usuarioId);
